renamed files, deleted old test files
